+++ luci-app-rakitanmanager/root/usr/bin/modem-orbit.py
@@ -1,5 +1,5 @@
-import sys
 import logging
+import sys
 from huawei_lte_api.Client import Client
 from huawei_lte_api.Connection import Connection
 import time
@@ -10,41 +10,6 @@ def get_wan_info(client):
     wan_ip_address = wan_info.get('WanIPAddress')
     device_name = wan_info.get('DeviceName')
     return wan_ip_address, device_name
-
-def main():
-    """Main function."""
-    if len(sys.argv) < 4:
-        print("Usage: python script.py <router_ip> <username> <password>")
-        sys.exit(1)
-    
-    router_ip = sys.argv[1]
-    username = sys.argv[2]
-    password = sys.argv[3]
-
-    connection_url = f'http://{username}:{password}@{router_ip}/'
-
-    with Connection(connection_url) as connection:
-        client = Client(connection)
-        
-        try:
-            if len(sys.argv) == 5 and sys.argv[4] == "ip":
-                wan_ip_address, _ = fetch_wan_info(client)
-                print(f"{wan_ip_address}")
-            else:
-                print_header("Auto Reconnect Modem Huawei", "@RTASERVER")
-                wan_ip_address, device_name = fetch_wan_info(client)
-                print_result("Device Name", device_name)
-                print_result("Current IP", wan_ip_address)
-                print("Initiating IP change process...")
-                initiate_ip_change(client)
-                time.sleep(5)  
-                print("Waiting for the IP to be changed...")
-                wan_ip_address_after_plmn, _ = fetch_wan_info(client)
-                print_result("New IP", wan_ip_address_after_plmn)
-                print_success("IP has been successfully changed.")
-        
-        except Exception as e:
-            print_error(f"An error occurred: {e}")
 
 def fetch_wan_info(client):
     """Fetch WAN IP address and device name."""
@@ -71,11 +36,43 @@ def print_result(label, value):
 
 def print_success(message):
     """Print success message."""
-    print("\n" + message)
+    print("\n\033[92m" + message + "\033[0m")
 
 def print_error(message):
     """Print error message."""
-    print("\n" + message)
+    print("\n\033[91m" + message + "\033[0m")
+
+def main():
+    """Main function."""
+    router_ip = sys.argv[1]
+    username = sys.argv[2]
+    password = sys.argv[3]
+
+    connection_url = f'http://{username}:{password}@{router_ip}/'
+
+    with Connection(connection_url) as connection:
+        client = Client(connection)
+        
+        try:
+            print_header("Auto Reconnect Modem Huawei", "@RTASERVER")           
+            
+            wan_ip_address, device_name = fetch_wan_info(client)
+            print_result("Device Name", device_name)
+            print_result("Current IP", wan_ip_address)
+          
+            print("Initiating IP change process...")
+            initiate_ip_change(client)
+            
+            time.sleep(5)  
+            
+            print("Waiting for the IP to be changed...")
+            wan_ip_address_after_plmn, _ = fetch_wan_info(client)
+            print_result("New IP", wan_ip_address_after_plmn)
+            
+            print_success("IP has been successfully changed.")
+        
+        except Exception as e:
+            print_error(f"An error occurred: {e}")
 
 if __name__ == "__main__":
     main()